+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2020-06-30
+### Changed
+- Consumer application to use same logging framework as producer (SLF4J)
+
+### Removed
+- log4j dependency from consumer application
+
 ## [1.1.0] - 2020-04-08
 ### Added
 - CHANGELOG, NOTICE, CONTRIBUTING, and CODE_OF_CONDUCT files